+++ src/test/java/org/opentripplanner/generate/doc/UpdaterConfigDocTest.java
@@ -50,7 +50,7 @@ public class UpdaterConfigDocTest {
   public void updateRouterConfigurationDoc() {
     NodeAdapter node = readBuildConfig();
 
-    // Read and close inout file (same as output file)
+    // Read and close input file (same as output file)
     String template = readFile(TEMPLATE);
     String original = readFile(OUT_FILE);